% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.ycevo}
\alias{augment.ycevo}
\title{Augment data with predicted discount function and yield curve}
\usage{
\method{augment}{ycevo}(x, newdata = NULL, loess = TRUE, ...)
}
\arguments{
\item{x}{A \link{ycevo} object}

\item{newdata}{A data frame containing time-to-maturity in years \code{tau} and
the quotation date at which the discount function and the yield curve is to
be predicted. The quotation date is to be named the same as the quotation
date column in \code{x}, depending on the name of the quotation date column when
it was first provided to the \code{data} argument \code{\link[=ycevo]{ycevo()}}.The default is
\code{qdate}.}

\item{loess}{Logical. Whether the returned discount function and yield curve
are loess smoothed.}

\item{...}{Additional arguments required for generic consistency. Currently
not used. Warning: A misspelled argument will not raise an error. The
misspelled argument will be either disregarded, or the default value will
be applied if one exists.}
}
\value{
\code{newdata} augmented with \code{.discount} and \code{.yield} for the discount
function and the yield curve respectively.
}
\description{
Augment data with predicted discount function and yield curve
}
\details{
If \code{newdata} is not provided, returns the discount function and yield curve
values at points specified for the estimation in \code{\link[=ycevo]{ycevo()}}.

If \code{newdata} is provided, the discount function at the time-to-maturities
specified in \code{newdata} will be generated from loess smoothing (see
\code{\link[stats:loess]{stats::loess()}}), then interpolated to produce the discount function at the
quotation date specified in \code{newdata}.
}
\examples{
# Simulating bond data
bonds <- ycevo_data(n = 10)
\donttest{
# Estimation can take up to 30 seconds
res <- ycevo(bonds, x = lubridate::ymd("2023-03-01"))
# Augmentation
augment(res)
}

}
\seealso{
\code{\link[=ycevo]{ycevo()}}
}
