% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_cf_slist}
\alias{calc_cf_slist}
\title{Create a list of sparse matrices to represent cash flows}
\usage{
calc_cf_slist(data)
}
\arguments{
\item{data}{data for bonds including quotation days, bond id,
time until payment and payment amount.}
}
\value{
If all the required information is present, then the
output will be a list with length equal to the number of quotation
days in the data. Each element of the output will be a sparseMatrix
object.
}
\description{
Tranforms data into a format suitable for estimation
}
\details{
This function converts and extracts coupon payments over
quotation days and different bonds from the raw data
into a list of sparse matrices for estimation
}
\examples{
\dontrun{
calc_cf_slist(USbonds)
}
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
