% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_day_idx}
\alias{calc_day_idx}
\title{Provide indices in relation to time grids}
\usage{
calc_day_idx(data, ugrid, hu)
}
\arguments{
\item{data}{a bond data frame. See \code{?USbonds} for an example data structure.}

\item{ugrid}{vector of quotation date grid}

\item{hu}{vector of quotation date bandwidth}
}
\value{
Matrix. The start and end of the qdates included in the ugrid kernel windows
}
\description{
This function provide indices for the start and end of the qdates included in the kernel windows for each \code{ugrid}.
}
\examples{
 ugrid <- c(0.2,0.4)
 hu <- c(0.18,0.18)
 out <- calc_day_idx(data = USbonds, ugrid = ugrid, hu = hu)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
