% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{calc_dbar}
\alias{calc_dbar}
\title{A component of the \code{estimate_yield} function that calculates \eqn{dbar}}
\source{
Koo, B., La Vecchia, D., & Linton, O. B. (2019). Estimation of a Nonparametric model for Bond Prices from Cross-section and Time series Information. Available at SSRN3341344.
}
\usage{
calc_dbar(
  data,
  ugrid,
  hu,
  rgrid,
  hr,
  xgrid,
  hx,
  price_slist,
  cf_slist,
  interest,
  units = 365
)
}
\arguments{
\item{data}{A data frame; bond data to estimate discount curve from.}

\item{ugrid}{A length T numeric vector; the times at which the discount curve will be estimated.}

\item{hu}{A length T numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time at which the discount curve is estimated,}

\item{rgrid}{(Optional) A length K numeric vector of interest rate grid values}

\item{hr}{(Optional) A length K numeric vector of interest rate grid bandwidths}

\item{xgrid}{A length m numeric vector, or either a 1 x x or T x x numeric matrix. If a T x x matrix, each row represents the time-to-maturity grid
that each time-to-maturity in the discount function is compared against. Otherwise the same time-to-maturity grid is repeated for each of the T ugrid values}

\item{hx}{An numeric vector of length T, or if xgrid is a matrix, a numeric matrix of the same dimensions as xgrid.
A vector hx for a T x m matrix xgrid repeats values for each row of qgrid.
bandwidth parameter determining the size of the window that corresponds to each time-to-maturity.}

\item{price_slist}{(Optional) A list of matrices, generated by calc_price_slist.}

\item{cf_slist}{(Optional) A list of matrices, generated by calc_cf_slist.}

\item{interest}{(Optional) A vector of daily short term interest rates}

\item{units}{(Optional) number of tupq per xgrid (e.g. 365 for daily data with annual grid values). Defaults to 365}
}
\value{
Data frame with the follwing variables

\describe{
  \item{ug}{Same as input \code{ugrid}}
  \item{dbar_numer}{Numerator in dbar. See \code{Source}}
  \item{dbar_denom}{Denominator in dbar. See \code{Source}}
  \item{xg}{Same as input \code{xgrid}}
}
}
\description{
Internal function that estimates the discount function without taking into account
cross products of coupon payments. Not to be used independently. Exported for documentation purpose.
}
\author{
Nathaniel Tomasetti
}
