% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_price_slist}
\alias{calc_price_slist}
\title{Create a bond price list from the data}
\usage{
calc_price_slist(data)
}
\arguments{
\item{data}{a bond data}
}
\value{
price_slist
}
\description{
Tranforms data into a format suitable for estimation
}
\details{
This function converts and extracts bond prices over
quotation days and different bonds from the raw data
into a list of sparse matrices for estimation
}
\examples{
calc_price_slist(USbonds)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
