% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_tupq_idx}
\alias{calc_tupq_idx}
\title{Provide indices in relation to time-to-maturity grids}
\usage{
calc_tupq_idx(data, xgrid, hx, units = 365)
}
\arguments{
\item{data}{Bond dataframe}

\item{xgrid}{vector of the time-to-maturity grid}

\item{hx}{vector of the time-to-maturity grid bandwidth}
}
\value{
Matrix. The first and last xgrid included in the current kernel window.
}
\description{
Provide indices in relation to time-to-maturity grids
}
\details{
This function provides indices for the first and last xgrid included in the current kernel window
}
\examples{
xgrid <- c(30, 60, 90) / 365
hx <- c(15, 15 , 15) / 365
out <- calc_tupq_idx(data = USbonds, xgrid = xgrid, hx = hx)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
