% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_ux_window}
\alias{calc_ux_window}
\title{Weights time to maturity grid}
\usage{
calc_ux_window(data, xgrid, hx, units = 365)
}
\arguments{
\item{data}{a bond data frame}

\item{xgrid}{vector of the time-to-maturity grid}

\item{hx}{vector of the time-to-maturity grid bandwidth}
}
\value{
Matrix with number of columns being the length of \code{xgrid} and number of rows being the number of unique qdates.
Each column represents the weights of each qdate for that \code{xgrid}.
Each column is a \code{xgrid} date with the weights of the qdates used in discount function estimation. qdates correspond to rows.
}
\description{
Apply kernel in relation to time-to-maturity grids
}
\details{
This function generates a weight function attached to each time grid
for the estimation of a discount function
}
\examples{
xgrid <- c(30, 60, 90) / 365
hx <- c(15, 15 , 15) / 365
out <- calc_ux_window(data = USbonds, xgrid = xgrid, hx = hx)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
