% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{create_xgrid_hx}
\alias{create_xgrid_hx}
\title{Automatic selection of xgrid and hx values}
\usage{
create_xgrid_hx(
  data,
  ugrid,
  hu,
  qgrid,
  hq,
  min_points,
  rgrid,
  hr,
  interest,
  units = 365
)
}
\arguments{
\item{data}{Bond dataframe. See \code{?USbonds} for an example data strcture..}

\item{ugrid}{A single value for ugrid between 0 and 1}

\item{hu}{A single value for the bandwidth of the ugrid value}

\item{qgrid}{vector of dense time to maturity grid}

\item{hq}{vector of dense time to maturity bandwidth}

\item{min_points}{Integer, minimum number of maturing bonds in a qgrid range to be included in xgrid.}

\item{rgrid}{Optional, a single value for rgrid}

\item{hr}{Optional, A single value for the bandwidth of the rgrid value}

\item{interest, }{Optional, a vector of daily interest rates for use with rgrid. Must have a length equal to the number of unique qdates in data}

\item{units, }{Optional, number of units per period. Eg 365 for daily data, 12 for monthly.
Grid values without maturing bonds do not have sufficient data for stable estimates.}
}
\value{
List of \code{xgrid}, \code{hx}, \code{qgrid} and \code{hq}. \code{qgrid} and \code{hq} is the same as input. 
For the usage of created \code{xgrid} and \code{hx}, see \code{\link{estimate_yield}}.
}
\description{
Selects xgrid and hx from values of qgrid and hq with a given number of maturing bonds
}
\details{
Automatically select values for sparse time to maturity \code{xgrid} and \code{hx}
for a given dense time to maturity value of qgrid and hq and quotation date ugrid and hu.
The length of the provided xgrid may change for different ugrid values, so it is recommended
that the function is called separately for different values of ugrid.
}
\examples{
 ugrid <- 0.2
 hu <- 0.18
 qgrid <- c(30, 60, 90) / 365
 hq <- c(15, 15 , 15) / 365
 out <- create_xgrid_hx(data = USbonds, ugrid = ugrid, hu = hu, qgrid = qgrid,hq =hq, min_points = 5)
}
\author{
Nathaniel Tomasetti
}
