% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{dhat_var}
\alias{dhat_var}
\title{Estimate variances for each point on the discount function and yield curve}
\source{
Koo, B., La Vecchia, D., & Linton, O. B. (2019). Estimation of a Nonparametric model for Bond Prices from Cross-section and Time series Information. Available at SSRN3341344.
}
\usage{
dhat_var(data, ugrid, hu, xgrid, hx, perrors, dhat, cf_slist)
}
\arguments{
\item{data}{A data frame; bond data to estimate discount curve from.}

\item{ugrid}{A numeric vector; the times at which the discount curve will be estimated.}

\item{hu}{A numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time at which the discount curve is estimated,}

\item{xgrid}{A numeric matrix, each row represents the time-to-maturity grid
for the discount function at the corresponding time.}

\item{hx}{A numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time-to-maturity.}

\item{perrors}{A dataframe with columns qdate, crspid and perror, where perror is the in-sample pricing error (residual) for the given crspid and qdate.}

\item{dhat}{Dataframe obtained from estimate_dhat function}

\item{cf_slist}{(Optional) A list of matrices, generated by \code{\link{calc_cf_slist}}.}
}
\value{
Data frame of the estimated variance in addition to the yield and discount rate at each combination of the provided grids.
\describe{
  \item{dhat_var}{Estimated variance of the discount rate}
  \item{yield_var}{Estimated variance of the yield}
  \item{discount}{Estimated discount rate}
  \item{ug}{Same as input \code{ugrid}}
  \item{xg}{Same as input \code{xgrid}}
  \item{yield}{Estimated yield}
}
}
\description{
Non-parametric estimation of variances for each point on the discount function and yield curve
}
\details{
See \code{Source} for detailed estimation method. See vignette for examples.
}
\author{
Nathaniel Tomasetti
}
