% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTB3}
\alias{DTB3}
\title{Daily interest rates from 4/1/1954 to 28/12/2018}
\format{A data frame with 16955 rows and 2 variables:
\describe{
\item{date}{vector of dates}
\item{rate}{daily annualised interest rate in percentages}
}}
\usage{
DTB3
}
\description{
A dataset containing the daily 3 Month US Treasury Bill rate
}
\keyword{datasets}
