% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{epaker}
\alias{epaker}
\title{epaker kernel function}
\usage{
epaker(x)
}
\arguments{
\item{x}{value to apply kernel to}
}
\description{
Kernel function for grid windows
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
