% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{interpolate_discount}
\alias{interpolate_discount}
\title{Interpolates the discount rates resulting from estimate_yield}
\usage{
interpolate_discount(data, yield, treasury)
}
\arguments{
\item{data}{Bond dataframe}

\item{yield}{Output of estimate_yield}

\item{treasury}{Optional, dataframe of daily treasury bill rates with date and rate columns}
}
\description{
Applies three dimensional interpolation to find discount rates suitable for provided data
}
\details{
Given a bond data object and the output of estimate_yield, this function interpolates over each grid
to obtain new values of the discount rate for arbitrary rgrid / ugrid and qgrid values.
}
\author{
Nathaniel Tomasetti
}
