% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_ugrid}
\alias{interpolate_ugrid}
\title{One dimensional interpolation over ugrid}
\usage{
interpolate_ugrid(x, u, ugrid, interp, xvals)
}
\arguments{
\item{x, }{the value to be interpolated}

\item{u, }{the value to be interpolated}

\item{ugrid, }{vector, the range of u values we have data for}

\item{interp, }{matrix, element i,j gives the dhat value of the i'th xval and j'th ugrid}

\item{xvals, }{vector, the unique values of xvals that are included in interp. x must appear in this vector.}
}
\description{
Given the inputs, this interpolates dhat at a given xgrid over the nearest ugrid values
}
