% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{num_points}
\alias{num_points}
\title{Calculates the number of payments in each qgrid}
\usage{
num_points(.data, ugrid, hu, qgrid, hq)
}
\arguments{
\item{.data}{A data frame; bond data to estimate discount curve from.}

\item{ugrid}{A single value for ugrid between 0 and 1}

\item{hu}{A single value for the bandwidth of the ugrid value}

\item{qgrid}{A numeric matrix, each row represents the time-to-maturity grid
for the discount function at the corresponding time.}

\item{hq}{A numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time-to-maturity.}
}
\description{
Calculates the number of payments in each qgrid
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
