% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/autoplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{autoplot}
\title{Objects exported from other packages}
\seealso{
\code{\link[=augment.ycevo]{augment.ycevo()}}

\code{\link[=autoplot.ycevo]{autoplot.ycevo()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

