% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_data}
\alias{vignette_data}
\title{Results of the calculation in the vignette}
\format{A list required to produce sample figures and table}
\usage{
vignette_data
}
\description{
Results of the calculation in the vignette
}
\keyword{datasets}
