% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_yield}
\alias{vignette_yield}
\title{Results of the estimate_yield loop in the vignette}
\format{A data frame output from estimate_yield}
\usage{
vignette_yield
}
\description{
Results of the estimate_yield loop in the vignette
}
\keyword{datasets}
