% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycevo-package.R
\docType{package}
\name{ycevo-package}
\alias{ycevo-package}
\alias{_PACKAGE}
\title{Nonparametric Estimation of the Yield Curve Evolution}
\description{
Nonparametric estimation of the discount rate and yield curve.
}
\references{
Koo, B., La Vecchia, D., & Linton, O. (2021). Estimation of a nonparametric model for bond prices from cross-section and time series information. Journal of Econometrics, 220(2), 562-588.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bonsook/ycevo}
  \item Report bugs at \url{https://github.com/bonsook/ycevo/issues}
}

}
\author{
Bonsoo Koo \email{bonsoo.koo@monash.edu}

Kai-Yang Goh \email{kai-yang.goh@monash.edu}

Nathaniel Tomasetti \email{nathaniel.tomasetti@gmail.com}

Yangzhuoran Fin Yang (Maintainer) \email{fin.yang@monash.edu}
}
\keyword{package}
