% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycevo.R
\docType{package}
\name{ycevo}
\alias{ycevo}
\title{Non-Parametric Estimation of the Yield Curve Evolution}
\description{
Non-parametric estimation of the discount rate and yield curve.
}
\author{
Bonsoo Koo <bonsoo.koo@monash.edu>,
Kai-Yang Goh <kai-yang.goh@monash.edu>
Nathaniel Tomasetti <nathaniel.tomasetti@gmail.com>
Yangzhuoran Yang <fin.yang@monash.edu>
}
\keyword{package}
