% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_day_idx}
\alias{calc_day_idx}
\title{Provide indices in relation to time grids}
\usage{
calc_day_idx(data, ugrid, hu)
}
\arguments{
\item{data}{a bond data frame}

\item{ugrid}{vector of quotation date grid}

\item{hu}{vector of quotation date bandwidth}
}
\value{
calc_day_idx
}
\description{
Provide indices in relation to time grids
}
\details{
This function provide indices for the start and end of the qdates included in the ugrid kernel windows
}
\examples{
ugrid <- c(0.2,0.4)
hu <- c(0.18,0.18)
calc_day_idx(data = USbonds, ugrid = ugrid, hu = hu)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
