% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{calc_hhat_num}
\alias{calc_hhat_num}
\title{A component of the \code{estimate_yield} function that calculates \eqn{hhat}}
\source{
Koo, B., La Vecchia, D., & Linton, O. B. (2019). Estimation of a Nonparametric model for Bond Prices from Cross-section and Time series Information. Available at SSRN3341344.
}
\usage{
calc_hhat_num(
  data,
  ugrid,
  hu,
  rgrid,
  hr,
  xgrid,
  hx,
  qgrid,
  hq,
  cf_slist,
  interest,
  units = 465
)
}
\arguments{
\item{data}{A data frame; bond data to estimate discount curve from.}

\item{ugrid}{A length T numeric vector; the times at which the discount curve will be estimated.}

\item{hu}{A length T numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time at which the discount curve is estimated,}

\item{rgrid}{(Optional) A length K numeric vector of interest rate grid values}

\item{hr}{(Optional) A length K numeric vector of interest rate grid bandwidths}

\item{xgrid}{A length m numeric vector, or either a 1 x m or T x m numeric matrix. If a T x m matrix, each row represents the time-to-maturity grid
for the discount function at the corresponding time. Otherwise the same time-to-maturity grid is repeated for each of the T ugrid values}

\item{hx}{An numeric vector of length T, or if xgrid is a matrix, a numeric matrix of the same dimensions as xgrid.
A vector hx for a T x m matrix qgrid repeats values for each row of xgrid.
bandwidth parameter determining the size of the window that corresponds to each time-to-maturity.}

\item{qgrid}{(Optional), A length m numeric vector, or either a 1 x m or T x m numeric matrix, matching xgrid input with m allowed to be different to M.
If a T x m matrix, each row represents the time-to-maturity grid that each time-to-maturity in the discount function is compared against.
Otherwise the same time-to-maturity grid is repeated for each of the T ugrid values.
If m = M, and each entry of qgrid is identical to xgrid, estimation is performed without interpolaton of the h-hat matrix.
If the entries are not identical each entry of each row of qgrid must be within the range from the smallest to largest value of the corresponding row of xgrid,
and linear interpolation of the h-hat matrix is performed.
If omitted, qgrid is set equal to xgrid.}

\item{hq}{(Optional) An numeric vector of length T, or if qgrid is a matrix, a numeric matrix of the same dimensions as qgrid.
A vector hq for a T x M matrix qgrid repeats values for each row of qgrid.
bandwidth parameter determining the size of the window that corresponds to each time-to-maturity.
If omitted, hq is set equal to hx}

\item{cf_slist}{(Optional) A list of matrices, generated by calc_cf_slist.}

\item{interest}{(Optional) A vector of daily short term interest rates}

\item{units}{(Optional) number of tupq per xgrid (e.g. 365 for daily data with annual grid values). Defaults to 365}
}
\value{
Data frame with the following variables

\describe{
  \item{hhat_numer}{Numerator in H hat. See \code{Source}}
  \item{ug}{Same as input \code{ugrid}}
  \item{xg}{Same as input \code{xgrid}}
  \item{qg}{Same as input \code{qgrid}}
}
}
\description{
Internal function that calculates coupon payment cross products. 
Not to be used independently. Exported for documentation purpose.
}
\author{
Nathaniel Tomasetti
}
