% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_price_slist}
\alias{calc_price_slist}
\title{Create a bond price list from the data}
\usage{
calc_price_slist(data)
}
\arguments{
\item{data}{a bond data. See \code{?USbonds} for an example data structure.}
}
\value{
If all the required information is present, then the
output will be a list with length equal to the number of quotation
days in the data. Each element of the output will be a sparseMatrix
object for a quotation day with the number of rows being the number of bonds 
and the number of columns being the maximum of time-to-maturity.
Each row is a bond with price indicated at the corresponding column (time).
}
\description{
Tranforms data into a format suitable for estimation
}
\details{
This function converts and extracts bond prices over
quotation days and different bonds from the raw data
into a list of sparse matrices for estimation
}
\examples{
\donttest{
price <- calc_price_slist(USbonds)
}
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
