% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_r_window}
\alias{calc_r_window}
\title{Weights interest rate grid}
\usage{
calc_r_window(interest, rgrid, hr)
}
\arguments{
\item{interest}{vector of daily interest rates}

\item{rgrid}{vector of interest rate grid}

\item{hr}{vector of interest rate bandwidth}
}
\value{
Matrix
}
\description{
Generate a kernel weight function in relation to interest rate grids
}
\details{
This function generates a weight function attached to each interest rate grid
for the estimation of a discount function
}
\author{
Bonsoo Koo, Kai-Yang Goh and Nathaniel Tomasetti
}
\keyword{internal}
