% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_uu_window}
\alias{calc_uu_window}
\title{Weights time grid}
\usage{
calc_uu_window(data, ugrid, hu)
}
\arguments{
\item{data}{a bond dataframe}

\item{ugrid}{vector of the quotation date grid}

\item{hu}{vector of quotation date bandwidth}
}
\value{
Matrix with number of columns being the length of \code{ugrid} and number of rows being the number of unique qdates.
Each column represents the weights of each qdate for that ugrid.
}
\description{
Generate a kernel weight function in relation to time grids
}
\details{
This function generates a weight function attached to each quotation date grid
for the estimation of a discount function
}
\examples{
 ugrid <- c(0.2,0.4)
 hu <- c(0.18,0.18)
 out <- calc_uu_window(data = USbonds, ugrid = ugrid,hu = hu)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
