% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{calc_ux_window}
\alias{calc_ux_window}
\title{Weights time to maturity grid}
\usage{
calc_ux_window(data, xgrid, hx, units = 365)
}
\arguments{
\item{data}{a bond data frame}

\item{xgrid}{vector of the time-to-maturity grid}

\item{hx}{vector of the time-to-maturity grid bandwidth}
}
\value{
Matrix. The weight attached to each time grid
for the estimation of a discount function
}
\description{
Apply kernel in relation to time-to-maturity grids
}
\details{
This function generates a weight function attached to each time grid
for the estimation of a discount function
}
\examples{
xgrid <- c(30, 60, 90) / 365
hx <- c(15, 15 , 15) / 365
out <- calc_ux_window(data = USbonds, xgrid = xgrid, hx = hx)
}
\author{
Bonsoo Koo and Kai-Yang Goh
}
\keyword{internal}
