% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_yield}
\alias{estimate_yield}
\title{Estimate yield function}
\source{
Koo, B., La Vecchia, D., & Linton, O. B. (2019). Estimation of a Nonparametric model for Bond Prices from Cross-section and Time series Information. Available at SSRN3341344.
}
\usage{
estimate_yield(
  data,
  ugrid,
  hu,
  rgrid,
  hr,
  xgrid,
  hx,
  qgrid,
  hq,
  price_slist,
  cf_slist,
  interest,
  units = 365,
  loess = TRUE
)
}
\arguments{
\item{data}{A data frame; bond data to estimate discount curve from. See \code{?USbonds} for an example bond data structure.}

\item{ugrid}{A length T numeric vector; the times at which the discount curve will be estimated.}

\item{hu}{A length T numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time at which the discount curve is estimated,}

\item{rgrid}{(Optional) A length K numeric vector of interest rate grid values}

\item{hr}{(Optional) A length K numeric vector of interest rate grid bandwidths}

\item{xgrid}{(Optional) A length m numeric vector, or either a 1 x m or T x m numeric matrix.
If a T x m matrix, each row represents the time-to-maturity grid for a given ugrid value, otherwise the same xgrid values are repeated for each ugrid value.
If m = M, and each entry of qgrid is identical to xgrid, estimation is performed without interpolaton of the h-hat matrix.
If the entries are not identical, the qgrid values must not lie outside the range given by the smallest to largest xgrid value
Linear interpolation of the h-hat matrix is then performed.
If omitted, xgrid is set equal to qgrid.}

\item{hx}{A numeric object in the same format as xgrid containing the kernel bandwidth for each xgrid value.
If xgrid is omitted, hx is set to equal hq.}

\item{qgrid}{A length m numeric vector, or either a 1 x m or T x m numeric matrix.
Represents the qgrid of T time-to-maturities values to estimate the yield for.
If this is a T x m matrix, each row represents the grid for a given ugrid value.
If this is a vector or T x 1 matrix, entries represents the qgrid, which will be repeated for each ugrid.
To be used for estimation of hhat. See \code{Details}.}

\item{hq}{A numeric object in the same format as qgrid containing the kernel bandwidth for each qgrid value.}

\item{price_slist}{(Optional) A list of matrices, generated by \code{\link{calc_price_slist}}.}

\item{cf_slist}{(Optional) A list of matrices, generated by \code{\link{calc_cf_slist}}.}

\item{interest}{(Optional) A vector of daily short term interest rates}

\item{units}{(Optional) number of tupq per xgrid (e.g. 365 for daily data with annual grid values). Defaults to 365}

\item{loess}{(Optional) Logical. Whether the output estimated discount and yield are to be smoothed using locally estimated scatterplot smoothing (LOESS)}
}
\value{
Data frame of the yield and discount rate at each combination of the provided grids.
\describe{
  \item{discount}{Estimated discount rate}
  \item{ug}{Same as input \code{ugrid}}
  \item{qg}{Same as input \code{qgrid}}
  \item{yield}{Estimated yield}
}
}
\description{
Non-parametric estimation of the discount function and transformation to yield at given dates, time-to-maturities, and interest rates.
}
\details{
Provides a data.frame of the yield and discount rate at each combination of the provided grids. 
See \code{Source} for detailed estimation method. The estimation has five major steps:
\enumerate{
\item Estimation of dbar, a component of the discount rate that ignores cross products.
\item Estimation of hhat, the cross product component of the discount rate, using both qgrid and xgrid.
\item Interpolation of hhat when the provided xgrid values are not equal to those found in qgrid.
\item Solving a linear equation to result in dhat, the discount rate.
\item Transformation of the discount rate into a yield.
}
}
\examples{
\donttest{
ugrid <- 0.207171315
hu <- 0.2
max_tumat <- max(USbonds$tumat)
qgrid <- c(seq(30, 6 * 30, 30),  # Monthly up to six months
           seq(240, 2 * 365, 60),  # Two months up to two years
           seq(720 + 90, 6 * 365, 90),  # Three months up to six years
           seq(2160 + 120, 20 * 365, 120),  # Four months up to 20 years
           #               seq(20 * 365 + 182, 30 * 365, 182)) / 365 # Six months up to 30 years
           seq(20 * 365 + 182, 30.6 * 365, 182)) / 365
# Cut qgrid so that there is one value greater than the maximum to maturity
qgrid <- qgrid[1:min(which(qgrid >= max_tumat / 365))]

laggap <- qgrid - dplyr::lag(qgrid)
leadgap <- dplyr::lead(qgrid) - qgrid
hq <- vapply(1:length(qgrid), function(x) max(laggap[x], leadgap[x], na.rm = TRUE), runif(1))
grids <- create_xgrid_hx(USbonds, ugrid, hu, qgrid, hq, 5)

price <- calc_price_slist(USbonds)
cf <- calc_cf_slist(USbonds)
dhat <- estimate_yield(data = USbonds, ugrid = ugrid, hu = hu, xgrid = grids$xgrid,
               hx=grids$hx, qgrid = qgrid,hq= hq, price_slist=price, cf_slist = cf)
}

}
\author{
Nathaniel Tomasetti
}
