% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_yield}
\alias{estimate_yield}
\title{Estimate yield function}
\usage{
estimate_yield(
  data,
  xgrid,
  hx,
  tau,
  ht,
  rgrid = NULL,
  hr = NULL,
  interest = NULL,
  loess = TRUE,
  price_slist = NULL,
  cf_slist = NULL
)
}
\arguments{
\item{data}{Data frame; bond data to estimate discount curve from. See \code{?USbonds} for an example bond data structure.}

\item{xgrid}{Numeric vector of values between 0 and 1. 
Time grids over the entire time horizon (percentile) of the data at which the discount curve is evaluated.}

\item{hx}{Numeric vector of values between 0 and 1. Bandwidth parameter determining the size of the window 
that corresponds to each time point (\code{xgrid}).}

\item{tau}{Numeric matrix that
represents time-to-maturities (column) in years for each time point \code{xgrid} (row) at which
the discount curve is evaluated. If there is only one time point \code{xgrid}, \code{tau} can be a numeric vector.
 See \code{Details}.}

\item{ht}{Numeric matrix that
represents bandwidth parameter determining the size of the window 
that corresponds to each time-to-maturities (\code{tau}). 
The same unit as \code{tau}.}

\item{rgrid}{(Optional) Numeric vector of interest rate grids in percentage at which the discount curve is evaluated, e.g. 4.03 means at interet rate of 4.03\%.}

\item{hr}{(Optional) Numeric vector of bandwidth parameter in percentage determining the size of the window
that corresponds to each interest rate grid (rgrid).}

\item{interest}{(Optional) Numeric vector of daily short term interest rates. 
The length is the same as the numebr of quotation dates included in the data, 
i.e. one interest rate per day.}

\item{loess}{(Optional) Logical. Whether the output estimated discount and yield are to be smoothed using locally estimated scatterplot smoothing (LOESS)}

\item{price_slist}{(Optional) A list of matrices, generated by \code{calc_price_slist}.}

\item{cf_slist}{(Optional) A list of matrices, generated by \code{calc_cf_slist}.}
}
\value{
Data frame of the yield and discount rate at each combination of the provided grids.
\describe{
  \item{discount}{Estimated discount rate}
  \item{xgrid}{Same as input \code{xgrid}}
  \item{tau}{Same as input \code{tau}}
  \item{yield}{Estimated yield}
}
}
\description{
Nonparametric estimation of discount functions at given dates, time-to-maturities, and interest rates and their transformation to the yield curves.
}
\details{
\eqn{d(\tau, X_t)} is the discount function at given covariates \eqn{X_t} (dates \code{xgrid} and 
interest rates \code{rgrid}), and given time-to-maturities \eqn{\tau} (\code{tau}).
\eqn{y(\tau, X_t)} is the yield curve at given covariates \eqn{X_t} (dates \code{xgrid} and 
interest rates \code{rgrid}), and given time-to-maturities \eqn{\tau} (\code{tau}).

We pursue the minimum of the following 
smoothed sample least squares objective function for any smooth function \eqn{d(.)}:
\deqn{Q(d) = \sum^T_{t=1}\sum^n_{i=1}\int\{p_{it}-\sum^{m_{it}}_{j=1}c_{it}(\tau_{ij})d(s_{ij}, x)\}^2 \sum^{m_{it}}_{k=1}\{K_h(s_{ik}-\tau_{ik})ds_{ik}\}K_h(x-X_t)dx,}
where a bond \eqn{i} has a price \eqn{p_i} at time t with a set of cash payments \eqn{c_1, c_2, \ldots, c_m} with a set of corresponding discount values
\eqn{d_1, d_2, \ldots, d_m},
\eqn{K_h(.) = K(./h)} is the kernel function with a bandwidth parameter \eqn{h},
the first kernel function is the kernel in space with bonds whose maturities \eqn{s_{ik}}
are close to the sequence \eqn{\tau_{ik}}, 
the second kernel function is the kernel in time and in interest rates with \eqn{x},  
which are close to the sequence \eqn{X_t}.
This means that bonds with similar cash flows, and traded in contiguous days, 
where the short term interest rates in the market are similar,
are combined for the estimation of the discount function at a point in space, in time, and in "interest rates".

The estimator for the discount function over time to maturity and time is
\deqn{\hat{d}=\arg\min_d Q(d).}
This function provides a data frame of the estimated yield and discount rate at each combination of the 
provided grids. The estimated yield is transformed from the estimated discount rate.

For more information on the estimation method, please refer to \code{References}.
}
\examples{
\donttest{
# USbonds provide data in year 2007
# Estimate the discount curve at the 20 percentile 
# (around mid March):
xgrid <- 0.2
hx <- 0.2
# Setting the time-to-maturity 
# at which the discount curve is estimated
# Beginning from 1 month, increasing by 1 month, up to 12 months 
tau <- seq(30, 360, 30) /365
ht <- rep(15, length(tau)) /365
# Estimation
dhat <- estimate_yield(data = USbonds, xgrid = xgrid,
                       hx = hx, tau = tau,ht= ht)
}

}
\references{
B. Koo, D. La Vecchia and O. Linton, Estimation of a nonparametric model for bond prices from cross-section and time series information. Journal of Econometrics (2020), https://doi.org/10.1016/j.jeconom.2020.04.014.
}
\author{
Nathaniel Tomasetti, Bonsoo Koo, and Yangzhuoran Yang
}
