% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_yield}
\alias{estimate_yield}
\title{Estimate yield function}
\usage{
estimate_yield(data, ugrid, hu, rgrid, hr, xgrid, hx, qgrid, hq,
  price_slist, cf_slist, interest, units = 365)
}
\arguments{
\item{data}{A data frame; bond data to estimate discount curve from.}

\item{ugrid}{A length T numeric vector; the times at which the discount curve will be estimated.}

\item{hu}{A length T numeric vector, bandwidth parameter determining the size of the window
that corresponds to each time at which the discount curve is estimated,}

\item{rgrid}{(Optional) A length K numeric vector of interest rate grid values}

\item{hr}{(Optional) A length K numeric vector of interest rate grid bandwidths}

\item{xgrid}{(Optional) A length m numeric vector, or either a 1 x m or T x m numeric matrix.
If a T x m matrix, each row represents the time-to-maturity grid for a given ugrid value, otherwise the same xgrid values are repeated for each ugrid value.
If m = M, and each entry of qgrid is identical to xgrid, estimation is performed without interpolaton of the h-hat matrix.
If the entries are not identical, the qgrid values must not lie outside the range given by the smallest to largest xgrid value
Linear interpolation of the h-hat matrix is then performed.
If omitted, xgrid is set equal to qgrid.}

\item{hx}{A numeric object in the same format as xgrid containing the kernel bandwidth for each xgrid value.
If xgrid is omitted, hx is set to equal hq.}

\item{qgrid}{A length m numeric vector, or either a 1 x m or T x m numeric matrix.
Represents the qgrid of T time-to-maturities values to estimate the yield for.
If this is a T x m matrix, each row represents the grid for a given ugrid value.
If this is a vector or T x 1 matrix, entries represents the qgrid, which will be repeated for each ugrid.}

\item{hq}{A numeric object in the same format as qgrid containing the kernel bandwidth for each qgrid value.}

\item{price_slist}{(Optional) A list of matrices, generated by calc_price_slist.}

\item{cf_slist}{(Optional) A list of matrices, generated by calc_cf_slist.}

\item{interest}{(Optional) A vector of daily short term interest rates}

\item{units}{(Optional) number of tupq per xgrid (e.g. 365 for daily data with annual grid values). Defaults to 365}
}
\description{
Estimation of discount function and transformation to yield at given dates, time to maturities, and interest rates.
}
\details{
Provides a data.frame of the yield and discount rate at each combination of the provided grids. Estimation follows (CITATION), with five major steps:
1. Estimation of dbar, a component of the discount rate that ignores cross products.
2. Estimation of hhat, the cross product component of the discount rate, using both qgrid and xgrid.
3. Interpolation of hhat when the provided xgrid values are not equal to those found in qgrid.
4. Solving a linear equation to result in dhat, the discount rate.
5. Transformation of the discount rate into a yield.
}
\examples{
\donttest{
ugrid<-0.2
hu<-0.2
xgrid<-seq(0,360,30) / 365
hq <-rep(15,length(xgrid)) / 365
price <- calc_price_slist(USbonds)
cf <- calc_cf_slist(USbonds)
estimate_yield(USbonds, ugrid, hu,xgrid =  xgrid,hq= hq, price_slist, cf_slist = cf)
}
}
\author{
Nathaniel Tomasetti
}
