% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycevo-package.R
\docType{package}
\name{ycevo-package}
\alias{ycevo-package}
\title{Nonparametric Estimation of the Yield Curve Evolution}
\description{
Nonparametric estimation of discount functions and yield curves.
}
\references{
Koo, B., La Vecchia, D., & Linton, O. (2021). Estimation of a nonparametric model for bond prices from cross-section and time series information. Journal of Econometrics, 220(2), 562-588.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bonsook/ycevo}
  \item Report bugs at \url{https://github.com/bonsook/ycevo/issues}
}

}
\author{
\strong{Maintainer}: Yangzhuoran Fin Yang \email{yangyangzhuoran@gmail.com} (\href{https://orcid.org/0000-0002-1232-8017}{ORCID})

Authors:
\itemize{
  \item Bonsoo Koo \email{Bonsoo.Koo@monash.edu}
}

Other contributors:
\itemize{
  \item Nathaniel Tomasetti [contributor]
  \item Kai-Yang Goh [contributor]
}

}
\keyword{package}
