% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycevo.R
\name{ycevo}
\alias{ycevo}
\title{Estimation of Yield Curve Dynamics}
\description{
This package estimates the dynamics of the yield curve using nonparametric
methods and tests for stationarity.
}
\details{
Package 'ycevo'
}
\author{
Bonsoo Koo <bonsoo.koo@monash.edu>,
Kai-Yang Goh <kai-yang.goh@monash.edu>
Nathaniel Tomasetti <nathaniel.tomasetti@gmail.com>
}
